Revert "added .gitignore"
This reverts commit 2b6b1682

+++ src/game/brickbraker/entities/fixed/FixedEntity.java
@@ -4,7 +4,7 @@ import game.brickbraker.Game;
 import game.brickbraker.entities.Entity;
 
 public abstract class FixedEntity extends Entity {
-    public FixedEntity(Game game, float x, float y, float width, float height) {
+    public FixedEntity(Game game, float x, float y, int width, int height) {
         super(game, x, y, width, height);
     }
 }
